backend/bikes/requirements.txt
Django==3.2.12
djangorestframework==3.14.0
PyJWT==2.6.0
